backend/services/docx_parser.py
import docx

def extract_text_from_docx(file_path: str)-> str:
    doc = docx.Document(file_path)
    full_text = []
    for para in doc.paragraphs:
        full_text.append(para.text)
    return "\n".join(full_text)